Add files via upload
These files are the rrd-databases. They can be regarded as a backup by the date of 01.JAN.2019 but basically they are more or less a reminder for myself, where to store them (in the RRA-Folder...) and to show how the folder structure looks like in the current weather station folder structure.